refactoring: feed, postpage, profile
