+++ Front/src/pages/Feed.jsx
@@ -1,501 +1,75 @@
-import React, { useEffect, useState, useRef } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { fetchPosts } from '../services/postsApi';
-import emojisData from '../assets/emojis.json';
+import React from 'react';
+import { Link } from 'react-router-dom';
 import LogoutModal from '../components/LogoutModal';
 import Toast from '../components/Toast';
-import '../styles/pages/Feed.css';
-import logo from '../assets/logo.png';
-import { Link } from 'react-router-dom';
-import { getCookie } from '../utils/cookies';
-import { addLike, removeLike, hasUserLikedPost } from '../services/likesApi';
 import EmojiPicker from '../components/EmojiPicker';
 import EditPostModal from '../components/EditPostModal';
 import { parseMentions } from '../utils/parseMentions.jsx';
-
-const MAX_CHARACTERS = 2048;
-const MAX_VISIBLE_LINES = 10;
+import { getCookie } from '../utils/cookies';
+import { useFeed, formatCount, getLineCount, getTruncatedText } from './hooks/useFeed';
+import '../styles/pages/Feed.css';
+import logo from '../assets/logo.png';
 
 function Feed({ logout }) {
-    const navigate = useNavigate();
-    const emojiBtnRef = useRef(null);
-    const [posts, setPosts] = useState([]);
-    const [loading, setLoading] = useState(true);
-    const [likedPosts, setLikedPosts] = useState({});
-    const [newPostText, setNewPostText] = useState('');
-    const [selectedImage, setSelectedImage] = useState(null);
-    const [showEmojiPicker, setShowEmojiPicker] = useState(false);
-    const [activeCategory, setActiveCategory] = useState(0);
-    const [expandedPosts, setExpandedPosts] = useState({});
-    const [showLogoutModal, setShowLogoutModal] = useState(false);
-    const [toast, setToast] = useState(null);
-    const [page, setPage] = useState(0);
-    const [pageSize] = useState(10);
-    const [hasMore, setHasMore] = useState(true);
-    const [loadingMore, setLoadingMore] = useState(false);
-    const [showMenuPostId, setShowMenuPostId] = useState(null);
-    const [showEditModal, setShowEditModal] = useState(false);
-    const [editingPost, setEditingPost] = useState(null);
-
-    const fileInputRef = useRef(null);
-    const textInputRef = useRef(null);
-
-    // Загрузка постов при монтировании
-    useEffect(() => {
-        loadPosts();
-    }, []);
-
-    const loadPosts = async () => {
-        setLoading(true);
-        try {
-            // Загружаем посты с бэкенда
-            const posts = await fetchPosts();
-
-            // Если постов нет — просто устанавливаем пустой массив (без моков!)
-            if (!posts || posts.length === 0) {
-                setPosts([]);
-                return;
-            }
-
-            // Загружаем статус лайков для текущего пользователя
-            const userData = getCookie('catsgram_user_data');
-            const currentUserId = userData ? JSON.parse(userData).id : null;
-
-            if (currentUserId) {
-                const likedStatuses = {};
-
-                await Promise.all(
-                    posts.map(async (post) => {
-                        try {
-                            const isLiked = await hasUserLikedPost(post.id, currentUserId);
-                            likedStatuses[post.id] = isLiked;
-                        } catch (err) {
-                            console.log(`Не удалось проверить лайк для поста ${post.id}`);
-                            likedStatuses[post.id] = false;
-                        }
-                    })
-                );
-
-                setLikedPosts(likedStatuses);
-            }
-
-            setPosts(posts);
-
-        } catch (error) {
-            console.error('Ошибка загрузки постов:', error);
-            // При ошибке — просто пустой массив, без моков
-            setPosts([]);
-            setToast({
-                message: 'Не удалось загрузить посты',
-                type: 'error'
-            });
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    // Обработка копирования ссылки
-    const handleCopyLink = async (postId) => {
-        const url = `${window.location.origin}/post/${postId}`;
-        try {
-            await navigator.clipboard.writeText(url);
-            setToast({ message: 'Ссылка скопирована!', type: 'success' });
-        } catch (err) {
-            setToast({ message: 'Не удалось скопировать ссылку', type: 'error' });
-        }
-        setShowMenuPostId(null);
-    };
-
-    // Обработка удаления поста
-    const handleDeletePost = async (postId) => {
-        if (!window.confirm('Вы уверены что хотите удалить пост?')) return;
-
-        try {
-            const response = await fetch(`http://localhost:8080/posts/${postId}`, {
-                method: 'DELETE',
-            });
-
-            if (!response.ok) throw new Error('Не удалось удалить пост');
-
-            setPosts(prev => prev.filter(p => p.id !== postId));
-            setToast({ message: 'Пост удалён', type: 'success' });
-        } catch (error) {
-            setToast({ message: 'Ошибка при удалении поста', type: 'error' });
-        }
-        setShowMenuPostId(null);
-    };
-
-    // Обработка редактирования
-    const handleEditPost = (post) => {
-        setShowEditModal(true);
-        setEditingPost(post);
-        setShowMenuPostId(null);
-    };
-
-    const loadMorePosts = async () => {
-        if (loadingMore || !hasMore) return;
-
-        setLoadingMore(true);
-        const nextPage = page + 1;
-
-        try {
-            const response = await fetch(
-                `http://localhost:8080/posts?from=${nextPage * pageSize}&size=${pageSize}&sort=desc`
-            );
-
-            if (!response.ok) throw new Error(`HTTP ${response.status}`);
-
-            const rawPosts = await response.json();
-
-            if (rawPosts.length < pageSize) {
-                setHasMore(false);
-            }
-
-            if (rawPosts.length === 0) {
-                setLoadingMore(false);
-                return;
-            }
-
-            const newPosts = await Promise.all(rawPosts.map(async (post) => {
-                let author = null;
-                try {
-                    const authorRes = await fetch(`http://localhost:8080/users/${post.authorId}`);
-                    if (authorRes.ok) author = await authorRes.json();
-                } catch {}
-
-                let imageUrl = null;
-                try {
-                    const imgRes = await fetch(`http://localhost:8080/posts/${post.id}/images`);
-                    if (imgRes.ok) {
-                        const images = await imgRes.json();
-                        if (images[0]?.id) {
-                            imageUrl = `http://localhost:8080/images/${images[0].id}`;
-                        }
-                    }
-                } catch {}
-
-                let likes = 0, comments = 0;
-                try {
-                    const [likesRes, commentsRes] = await Promise.all([
-                        fetch(`http://localhost:8080/posts/${post.id}/likes`),
-                        fetch(`http://localhost:8080/posts/${post.id}/comments`),
-                    ]);
-                    if (likesRes.ok) likes = (await likesRes.json()).length;
-                    if (commentsRes.ok) comments = (await commentsRes.json()).length;
-                } catch {}
-
-                return {
-                    id: post.id,
-                    authorId: post.authorId,
-                    text: post.description,
-                    createdAt: post.postDate,
-                    time: formatPostDate(post.postDate),
-                    username: author?.username || `User${post.authorId}`,
-                    avatar: author?.avatar || null,
-                    verified: author?.verified || false,
-                    likes,
-                    comments,
-                    imageUrl,
-                };
-            }));
-
-            setPosts(prev => [...prev, ...newPosts]);
-            setPage(nextPage);
-
-            const userData = getCookie('catsgram_user_data');
-            const currentUserId = userData ? JSON.parse(userData).id : null;
-
-            if (currentUserId) {
-                const newLikedStatuses = {};
-                await Promise.all(
-                    newPosts.map(async (post) => {
-                        try {
-                            const isLiked = await hasUserLikedPost(post.id, currentUserId);
-                            newLikedStatuses[post.id] = isLiked;
-                        } catch {}
-                    })
-                );
-                setLikedPosts(prev => ({ ...prev, ...newLikedStatuses }));
-            }
-
-        } catch (error) {
-            console.error('Ошибка загрузки дополнительных постов:', error);
-        } finally {
-            setLoadingMore(false);
-        }
-    };
-
-    React.useEffect(() => {
-        const handleClickOutside = (e) => {
-            const isInsideContainer = e.target.closest('.emoji-picker-container');
-            const isInsidePicker = e.target.closest('.emoji-picker-portal');
-
-            if (!isInsideContainer && !isInsidePicker) {
-                setShowEmojiPicker(false);
-            }
-            if (!e.target.closest('.menu-container')) {
-                setShowMenuPostId(null);
-            }
-        };
-        document.addEventListener('click', handleClickOutside);
-        return () => document.removeEventListener('click', handleClickOutside);
-    }, []);
-
-    useEffect(() => {
-        const handleScroll = () => {
-            const scrollTop = window.innerHeight + window.scrollY;
-            const documentHeight = document.documentElement.offsetHeight;
-            const threshold = 300;
-
-            if (scrollTop >= documentHeight - threshold && hasMore && !loadingMore) {
-                loadMorePosts();
-            }
-        };
-
-        window.addEventListener('scroll', handleScroll);
-        return () => window.removeEventListener('scroll', handleScroll);
-    }, [page, hasMore, loadingMore]);
-
-    useEffect(() => {
-        if (textInputRef.current) {
-            textInputRef.current.style.height = 'auto';
-            textInputRef.current.style.height = Math.min(
-                textInputRef.current.scrollHeight,
-                150
-            ) + 'px';
-        }
-    }, [newPostText]);
-
-    const handleAttachmentClick = () => {
-        fileInputRef.current?.click();
-    };
-
-    const handleLikeToggle = async (postId, currentLikes) => {
-        const userData = getCookie('catsgram_user_data');
-        const currentUserId = userData ? JSON.parse(userData).id : null;
-
-        if (!currentUserId) {
-            setToast({ message: 'Необходимо войти для лайков', type: 'error' });
-            return;
-        }
-
-        const isCurrentlyLiked = likedPosts[postId] || false;
-
-        try {
-            if (isCurrentlyLiked) {
-                await removeLike(postId, currentUserId);
-                setLikedPosts(prev => ({ ...prev, [postId]: false }));
-                setPosts(prev => prev.map(post =>
-                    post.id === postId ? { ...post, likes: Math.max(0, (post.likes || 0) - 1) } : post
-                ));
-            } else {
-                await addLike(postId, currentUserId);
-                setLikedPosts(prev => ({ ...prev, [postId]: true }));
-                setPosts(prev => prev.map(post =>
-                    post.id === postId ? { ...post, likes: (post.likes || 0) + 1 } : post
-                ));
-            }
-        } catch (error) {
-            console.error('Ошибка при переключении лайка:', error);
-            setToast({ message: 'Не удалось обновить лайк', type: 'error' });
-        }
-    };
-
-    const handleFileChange = (e) => {
-        const file = e.target.files?.[0];
-        if (file) {
-            const imageUrl = URL.createObjectURL(file);
-            setSelectedImage({
-                file,
-                preview: imageUrl,
-                name: file.name,
-            });
-        }
-    };
-
-    const handleRemoveImage = () => {
-        if (selectedImage?.preview) {
-            URL.revokeObjectURL(selectedImage.preview);
-        }
-        setSelectedImage(null);
-        if (fileInputRef.current) {
-            fileInputRef.current.value = '';
-        }
-    };
-
-    const handleEmojiClick = (emoji) => {
-        const textarea = textInputRef.current;
-        const remainingChars = MAX_CHARACTERS - newPostText.length;
-
-        if (remainingChars < emoji.length) return;
-
-        if (textarea) {
-            const start = textarea.selectionStart;
-            const end = textarea.selectionEnd;
-            const newText = newPostText.substring(0, start) + emoji + newPostText.substring(end);
-
-            if (newText.length <= MAX_CHARACTERS) {
-                setNewPostText(newText);
-                setTimeout(() => {
-                    textarea.focus();
-                    textarea.setSelectionRange(start + emoji.length, start + emoji.length);
-                }, 0);
-            }
-        } else {
-            if ((newPostText + emoji).length <= MAX_CHARACTERS) {
-                setNewPostText(newPostText + emoji);
-            }
-        }
-        setShowEmojiPicker(false);
-    };
-
-    const handleTextChange = (e) => {
-        const text = e.target.value;
-        if (text.length <= MAX_CHARACTERS) {
-            setNewPostText(text);
-        }
-    };
-
-    const handleKeyDown = (e) => {
-        if (e.key === 'Enter' && e.shiftKey) {
-            e.preventDefault();
-            if (newPostText.trim() || selectedImage) {
-                handlePublish();
-            }
-        }
-    };
-
-    const formatPostDate = (isoString) => {
-        if (!isoString) return '3 дн.';
-        const date = new Date(isoString);
-        const now = new Date();
-        const diffMs = now - date;
-        const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-        const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-
-        if (diffHours < 1) return 'Только что';
-        if (diffHours < 24) return `${diffHours} ч.`;
-        if (diffDays === 1) return 'Вчера';
-        if (diffDays < 7) return `${diffDays} дн.`;
-        return date.toLocaleDateString('ru-RU', { day: 'numeric', month: 'short' });
-    };
-
-    const handlePublish = async () => {
-        if (!newPostText.trim() && !selectedImage) {
-            setToast({ message: 'Напишите что-нибудь или добавьте фото', type: 'error' });
-            return;
-        }
-
-        const userData = getCookie('catsgram_user_data');
-        const currentUserId = userData ? JSON.parse(userData).id : null;
-
-        if (!currentUserId) {
-            setToast({ message: 'Необходимо войти для публикации', type: 'error' });
-            return;
-        }
-
-        try {
-            const response = await fetch('http://localhost:8080/posts', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({
-                    authorId: currentUserId,
-                    description: newPostText,
-                }),
-            });
-
-            if (!response.ok) {
-                const errorText = await response.text();
-                throw new Error(`HTTP ${response.status}: ${errorText}`);
-            }
-
-            const newPost = await response.json();
-
-            let authorData = null;
-            try {
-                const authorRes = await fetch(`http://localhost:8080/users/${currentUserId}`);
-                if (authorRes.ok) authorData = await authorRes.json();
-            } catch (e) {
-                console.log('Не удалось загрузить автора');
-            }
-
-            let imageUrl = null;
-            if (selectedImage?.file) {
-                try {
-                    const formData = new FormData();
-                    formData.append('image', selectedImage.file);
-                    const imgRes = await fetch(`http://localhost:8080/posts/${newPost.id}/images`, {
-                        method: 'POST',
-                        body: formData,
-                    });
-                    if (imgRes.ok) {
-                        const images = await imgRes.json();
-                        if (images[0]?.id) {
-                            imageUrl = `http://localhost:8080/images/${images[0].id}`;
-                        }
-                    }
-                } catch (imgError) {
-                    console.warn('Не удалось загрузить изображение');
-                }
-            }
-
-            const postForFeed = {
-                id: newPost.id,
-                authorId: newPost.authorId,
-                text: newPost.description,
-                createdAt: newPost.postDate,
-                time: formatPostDate(newPost.postDate),
-                username: authorData?.username || `User${currentUserId}`,
-                avatar: authorData?.avatar || null,
-                verified: authorData?.verified || false,
-                likes: 0,
-                comments: 0,
-                imageUrl: imageUrl,
-            };
-
-            setPosts(prev => [postForFeed, ...prev]);
-            setNewPostText('');
-            handleRemoveImage();
-            setToast({ message: 'Пост опубликован! 🎉', type: 'success' });
-
-        } catch (error) {
-            console.error('Ошибка публикации:', error);
-            setToast({ message: 'Не удалось опубликовать пост', type: 'error' });
-            // Больше не создаём локальные посты-заглушки
-        }
-    };
-
-    const togglePostExpand = (postId) => {
-        setExpandedPosts(prev => ({
-            ...prev,
-            [postId]: !prev[postId]
-        }));
-    };
-
-    const getLineCount = (text) => text ? text.split('\n').length : 0;
-
-    const getTruncatedText = (text, maxLines) => {
-        if (!text) return '';
-        const lines = text.split('\n');
-        return lines.length <= maxLines ? text : lines.slice(0, maxLines).join('\n');
-    };
-
-    const handleLogoutClick = () => setShowLogoutModal(true);
-
-    const handleLogoutConfirm = () => {
-        if (logout) logout();
-        setShowLogoutModal(false);
-        window.location.href = '/login';
-    };
-
-    const handleLogoutCancel = () => setShowLogoutModal(false);
-
-    const characterCount = newPostText.length;
-    const remainingChars = MAX_CHARACTERS - characterCount;
-    const isLimitReached = remainingChars === 0;
-    const isNearLimit = remainingChars <= 50;
-
-
+    const {
+        // Рефы
+        emojiBtnRef,
+        fileInputRef,
+        textInputRef,
+
+        // Данные постов
+        posts,
+        loading,
+        likedPosts,
+        expandedPosts,
+        hasMore,
+        loadingMore,
+
+        // Новый пост
+        newPostText,
+        selectedImage,
+        remainingChars,
+        isLimitReached,
+        isNearLimit,
+
+        // UI
+        showEmojiPicker,
+        setShowEmojiPicker,
+        showLogoutModal,
+        toast,
+        setToast,
+        showMenuPostId,
+        setShowMenuPostId,
+        showEditModal,
+        editingPost,
+
+        // Хелперы
+        MAX_VISIBLE_LINES,
+
+        // Обработчики
+        handlePublish,
+        handleLikeToggle,
+        handleDeletePost,
+        handleEditPost,
+        handleEditUpdate,
+        handleEditClose,
+        handleCopyLink,
+        handleFileChange,
+        handleRemoveImage,
+        handleAttachmentClick,
+        handleTextChange,
+        handleKeyDown,
+        handleEmojiClick,
+        togglePostExpand,
+        handleLogoutClick,
+        handleLogoutCancel,
+        handleLogoutConfirm,
+
+        // Утилиты
+        navigate,
+    } = useFeed(logout);
+
+    const currentUsername = JSON.parse(getCookie('catsgram_user_data'))?.username || 'user';
 
     return (
         <div className="feed-page">
@@ -510,10 +84,7 @@ function Feed({ logout }) {
                             <span className="nav-icon">📰</span>
                             <span>Лента</span>
                         </Link>
-                        <Link
-                            to={`/u/${JSON.parse(getCookie('catsgram_user_data'))?.username || 'user'}`}
-                            className="nav-item"
-                        >
+                        <Link to={`/u/${currentUsername}`} className="nav-item">
                             <span className="nav-icon">👤</span>
                             <span>Профиль</span>
                         </Link>
@@ -532,6 +103,7 @@ function Feed({ logout }) {
                 </aside>
 
                 <main className="feed-main">
+                    {/* ── Форма нового поста ── */}
                     <div className="create-post">
                         <div className="create-post-header">
                             <div className="user-avatar">U</div>
@@ -608,6 +180,7 @@ function Feed({ logout }) {
                         </div>
                     </div>
 
+                    {/* ── Лента постов ── */}
                     <div className="posts-feed">
                         {posts.length === 0 && !loading ? (
                             <div className="no-posts">
@@ -617,24 +190,21 @@ function Feed({ logout }) {
                         ) : (
                             posts.map((post) => {
                                 const isExpanded = expandedPosts[post.id] || false;
-                                const lineCount = getLineCount(post.text);
-                                const needsTruncate = lineCount > MAX_VISIBLE_LINES;
+                                const needsTruncate = getLineCount(post.text) > MAX_VISIBLE_LINES;
                                 const displayText = needsTruncate && !isExpanded
                                     ? getTruncatedText(post.text, MAX_VISIBLE_LINES)
                                     : post.text;
-                                const showReadMore = needsTruncate && !isExpanded;
-                                const showLess = needsTruncate && isExpanded;
 
                                 return (
                                     <article key={post.id} className="post-card">
+                                        {/* Шапка поста */}
                                         <div className="post-header">
                                             <div className="post-author">
                                                 <div className="post-avatar">
-                                                    {post.avatar ? (
-                                                        <img src={post.avatar} alt="avatar" />
-                                                    ) : (
-                                                        <span>{post.username?.[0]?.toUpperCase() || '👤'}</span>
-                                                    )}
+                                                    {post.avatar
+                                                        ? <img src={post.avatar} alt="avatar" />
+                                                        : <span>{post.username?.[0]?.toUpperCase() || '👤'}</span>
+                                                    }
                                                 </div>
                                                 <div className="post-author-info">
                                                     <span className="post-username">
@@ -644,6 +214,8 @@ function Feed({ logout }) {
                                                     <span className="post-time">{post.time || '3 дн.'}</span>
                                                 </div>
                                             </div>
+
+                                            {/* Меню поста */}
                                             <div className="menu-container">
                                                 <button
                                                     className="post-menu-btn"
@@ -659,7 +231,7 @@ function Feed({ logout }) {
                                                         <button className="menu-item" onClick={() => handleCopyLink(post.id)}>
                                                             🔗 Копировать ссылку
                                                         </button>
-                                                        {post.username === JSON.parse(getCookie('catsgram_user_data'))?.username && (
+                                                        {post.username === currentUsername && (
                                                             <>
                                                                 <button className="menu-item" onClick={() => handleEditPost(post)}>
                                                                     ✏️ Редактировать
@@ -674,53 +246,46 @@ function Feed({ logout }) {
                                             </div>
                                         </div>
 
-                                        {showEditModal && editingPost && (
+                                        {/* Модалка редактирования */}
+                                        {showEditModal && editingPost?.id === post.id && (
                                             <EditPostModal
                                                 post={editingPost}
-                                                onClose={() => {
-                                                    setShowEditModal(false);
-                                                    setEditingPost(null);
-                                                }}
-                                                onUpdate={(updatedData) => {
-                                                    setPosts(prev => prev.map(p =>
-                                                        p.id === editingPost.id ? { ...p, ...updatedData } : p
-                                                    ));
-                                                    setShowEditModal(false);
-                                                    setEditingPost(null);
-                                                    setToast({ message: 'Пост обновлён!', type: 'success' });
-                                                }}
+                                                onClose={handleEditClose}
+                                                onUpdate={handleEditUpdate}
                                             />
                                         )}
 
+                                        {/* Текст поста */}
                                         {displayText && (
                                             <div className="post-text">
                                                 {parseMentions(displayText)}
                                             </div>
                                         )}
 
-                                        {showReadMore && (
+                                        {needsTruncate && (
                                             <button className="read-more-btn" onClick={() => togglePostExpand(post.id)}>
-                                                Читать дальше ↓
-                                            </button>
-                                        )}
-                                        {showLess && (
-                                            <button className="read-more-btn" onClick={() => togglePostExpand(post.id)}>
-                                                Свернуть ↑
+                                                {isExpanded ? 'Свернуть ↑' : 'Читать дальше ↓'}
                                             </button>
                                         )}
 
+                                        {/* Изображение поста */}
                                         {post.imageUrl && (
                                             <div className="post-image-container">
-                                                <img src={post.imageUrl} alt="post" className="post-image"
-                                                     onError={(e) => { e.target.style.display = 'none'; }} />
+                                                <img
+                                                    src={post.imageUrl}
+                                                    alt="post"
+                                                    className="post-image"
+                                                    onError={(e) => { e.target.style.display = 'none'; }}
+                                                />
                                             </div>
                                         )}
 
+                                        {/* Статистика */}
                                         <div className="post-stats">
                                             <div className="post-stats-left">
                                                 <button
                                                     className={`stat-btn like-btn ${likedPosts[post.id] ? 'liked' : ''}`}
-                                                    onClick={() => handleLikeToggle(post.id, post.likes)}
+                                                    onClick={() => handleLikeToggle(post.id)}
                                                 >
                                                     <span className="like-icon">{likedPosts[post.id] ? '❤️' : '🤍'}</span>
                                                     <span>{formatCount(post.likes || 0)}</span>
@@ -766,12 +331,6 @@ function Feed({ logout }) {
             )}
         </div>
     );
-}
-
-function formatCount(num) {
-    if (num >= 1000000) return (num / 1000000).toFixed(1) + 'M';
-    if (num >= 1000) return (num / 1000).toFixed(1) + 'K';
-    return num.toString();
 }
 
 export default Feed;
