+++ Front/src/pages/PostPage.jsx
@@ -1,316 +1,53 @@
-import React, { useEffect, useState, useRef } from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
-import { getCookie } from '../utils/cookies';
-import { addComment, getPostComments, getCommentAuthor } from '../services/commentsApi';
-import { addLike, removeLike, hasUserLikedPost } from '../services/likesApi';
-import { updatePost, deletePost } from '../services/postsApi';
+import React from 'react';
 import Layout from '../components/Layout';
 import Toast from '../components/Toast';
 import EditPostModal from '../components/EditPostModal';
-import '../styles/pages/PostPage.css';
 import EmojiPicker from '../components/EmojiPicker';
-import {parseMentions} from "../utils/parseMentions.jsx";
-
+import { parseMentions } from '../utils/parseMentions.jsx';
+import { usePostPage } from './hooks/usePostPage';
+import '../styles/pages/PostPage.css';
 
 function PostPage() {
-    const { postId } = useParams();
-    const navigate = useNavigate();
+    const {
+        // Рефы
+        menuRef,
+        commentInputRef,
+        commentEmojiBtnRef,
 
-    const [post, setPost] = useState(null);
-    const [comments, setComments] = useState([]);
-    const [commentText, setCommentText] = useState('');
-    const [loading, setLoading] = useState(true);
-    const [submitting, setSubmitting] = useState(false);
-    const [liked, setLiked] = useState(false);
-    const [toast, setToast] = useState(null);
-    const [showMenu, setShowMenu] = useState(false);
-    const [showEditModal, setShowEditModal] = useState(false);
-    const [showEmojiPicker, setShowEmojiPicker] = useState(false);
-    const [activeEmojiCategory, setActiveEmojiCategory] = useState(0);
-    const [canGoBack, setCanGoBack] = useState(false);
-    const [isOwner, setIsOwner] = useState(false);
-    const [sortNewest, setSortNewest] = useState(true); // true = новые, false = старые
+        // Данные
+        post,
+        comments,
+        commentText,
+        setCommentText,
+        loading,
+        submitting,
+        liked,
+        isOwner,
+        sortNewest,
 
-    const menuRef = useRef(null);
-    const commentInputRef = useRef(null);
-    const commentEmojiBtnRef = useRef(null);
+        // UI
+        showMenu,
+        setShowMenu,
+        showEditModal,
+        setShowEditModal,
+        showEmojiPicker,
+        setShowEmojiPicker,
+        toast,
+        setToast,
 
-    useEffect(() => {
-        setCanGoBack(window.history.length > 1);
-        loadPostData();
+        // Обработчики
+        handleLikeToggle,
+        handleCommentSubmit,
+        handleEmojiClick,
+        handleBack,
+        handleCopyLink,
+        handleEditPost,
+        handleDeletePost,
+        handleSortToggle,
 
-        const handleClickOutside = (e) => {
-            if (menuRef.current && !menuRef.current.contains(e.target)) {
-                setShowMenu(false);
-            }
-            if (!e.target.closest('.emoji-picker-container')) {
-                setShowEmojiPicker(false);
-            }
-        };
-        document.addEventListener('click', handleClickOutside);
-        return () => document.removeEventListener('click', handleClickOutside);
-    }, [postId]);
-
-    const loadPostData = async () => {
-        setLoading(true);
-        try {
-            const userData = getCookie('catsgram_user_data');
-            const currentUserId = userData ? JSON.parse(userData).id : null;
-
-            const postResponse = await fetch(`http://localhost:8080/posts/${postId}`);
-            if (!postResponse.ok) throw new Error('Пост не найден');
-            const postData = await postResponse.json();
-
-            const author = await fetchAuthor(postData.authorId);
-            const images = await fetchPostImages(postData.id);
-            const likes = await fetchPostLikes(postData.id);
-            const isLiked = currentUserId ? await hasUserLikedPost(postData.id, currentUserId) : false;
-
-            setIsOwner(currentUserId === postData.authorId);
-
-            const rawComments = await getPostComments(postData.id);
-            const commentsWithAuthors = await Promise.all(
-                rawComments.map(async (comment) => {
-                    const commentAuthor = await getCommentAuthor(comment.authorId);
-                    return {
-                        ...comment,
-                        author: commentAuthor,
-                        time: formatCommentDate(comment.createdAt),
-                    };
-                })
-            );
-
-            // Сортировка по умолчанию - новые сверху
-            sortComments(commentsWithAuthors, true);
-
-            setPost({
-                id: postData.id,
-                authorId: postData.authorId,
-                text: postData.description || '',
-                createdAt: postData.postDate,
-                time: formatPostDate(postData.postDate),
-                username: author?.username || `User${postData.authorId}`,
-                avatar: author?.avatar || null,
-                verified: author?.verified || false,
-                likes: likes.length,
-                imageUrl: images[0]?.url || null,
-                hasImage: images.length > 0,
-            });
-
-            setLiked(isLiked);
-
-        } catch (error) {
-            console.error('Ошибка загрузки поста:', error);
-            setToast({ message: 'Не удалось загрузить пост', type: 'error' });
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    const sortComments = (commentsList, newest) => {
-        const sorted = [...commentsList].sort((a, b) => {
-            if (newest) {
-                return new Date(b.createdAt) - new Date(a.createdAt);
-            } else {
-                return new Date(a.createdAt) - new Date(b.createdAt);
-            }
-        });
-        setComments(sorted);
-    };
-
-    const handleSortToggle = () => {
-        const newSort = !sortNewest;
-        setSortNewest(newSort);
-        sortComments(comments, newSort);
-    };
-
-    const fetchAuthor = async (authorId) => {
-        try {
-            const response = await fetch(`http://localhost:8080/users/${authorId}`);
-            if (!response.ok) return null;
-            return await response.json();
-        } catch { return null; }
-    };
-
-    const fetchPostLikes = async (postId) => {
-        try {
-            const response = await fetch(`http://localhost:8080/posts/${postId}/likes`);
-            if (!response.ok) return [];
-            return await response.json();
-        } catch { return []; }
-    };
-
-    const fetchPostImages = async (postId) => {
-        try {
-            const response = await fetch(`http://localhost:8080/posts/${postId}/images`);
-            if (!response.ok) return [];
-            const images = await response.json();
-            return images.map(img => ({ id: img.id, url: `http://localhost:8080/images/${img.id}` }));
-        } catch { return []; }
-    };
-
-    const formatPostDate = (isoString) => {
-        if (!isoString) return '3 дн.';
-        const date = new Date(isoString);
-        const now = new Date();
-        const diffMs = now - date;
-        const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-        const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-        if (diffHours < 1) return 'Только что';
-        if (diffHours < 24) return `${diffHours} ч.`;
-        if (diffDays === 1) return 'Вчера';
-        if (diffDays < 7) return `${diffDays} дн.`;
-        return date.toLocaleDateString('ru-RU', { day: 'numeric', month: 'long' });
-    };
-
-    const formatCommentDate = (isoString) => {
-        if (!isoString) return '';
-        const date = new Date(isoString);
-        const now = new Date();
-        const diffMs = now - date;
-        const diffMinutes = Math.floor(diffMs / (1000 * 60));
-        const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-        const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-        if (diffMinutes < 1) return 'Только что';
-        if (diffMinutes < 60) return `${diffMinutes} мин.`;
-        if (diffHours < 24) return `${diffHours} ч.`;
-        if (diffDays === 1) return 'Вчера';
-        if (diffDays < 7) return `${diffDays} дн.`;
-        return date.toLocaleDateString('ru-RU', { day: 'numeric', month: 'short' });
-    };
-
-    const formatCount = (num) => {
-        if (num >= 1000000) return (num / 1000000).toFixed(1) + 'M';
-        if (num >= 1000) return (num / 1000).toFixed(1) + 'K';
-        return num.toString();
-    };
-
-    const handleLikeToggle = async () => {
-        const userData = getCookie('catsgram_user_data');
-        const currentUserId = userData ? JSON.parse(userData).id : null;
-        if (!currentUserId) {
-            setToast({ message: 'Необходимо войти для лайков', type: 'error' });
-            return;
-        }
-        try {
-            if (liked) {
-                await removeLike(postId, currentUserId);
-                setLiked(false);
-                setPost(prev => ({ ...prev, likes: Math.max(0, prev.likes - 1) }));
-            } else {
-                await addLike(postId, currentUserId);
-                setLiked(true);
-                setPost(prev => ({ ...prev, likes: prev.likes + 1 }));
-            }
-        } catch (error) {
-            setToast({ message: 'Не удалось обновить лайк', type: 'error' });
-        }
-    };
-
-    const handleCommentSubmit = async (e) => {
-        e.preventDefault();
-        if (!commentText.trim()) return;
-
-        const userData = getCookie('catsgram_user_data');
-        const currentUserId = userData ? JSON.parse(userData).id : null;
-        if (!currentUserId) {
-            setToast({ message: 'Необходимо войти для комментариев', type: 'error' });
-            return;
-        }
-
-        setSubmitting(true);
-        try {
-            const newComment = await addComment(postId, currentUserId, commentText);
-            const author = await fetchAuthor(currentUserId);
-            const commentWithAuthor = {
-                ...newComment,
-                author,
-                time: 'Только что',
-            };
-
-            // Добавляем комментарий и сортируем
-            const updatedComments = sortNewest
-                ? [commentWithAuthor, ...comments]
-                : [...comments, commentWithAuthor];
-
-            setComments(updatedComments);
-            setCommentText('');
-            setToast({ message: 'Комментарий добавлен!', type: 'success' });
-        } catch (error) {
-            setToast({ message: 'Не удалось добавить комментарий', type: 'error' });
-        } finally {
-            setSubmitting(false);
-        }
-    };
-
-    const handleEmojiClick = (emoji) => {
-        const textarea = commentInputRef.current;
-        if (textarea) {
-            const start = textarea.selectionStart;
-            const end = textarea.selectionEnd;
-            const newText = commentText.substring(0, start) + emoji + commentText.substring(end);
-            setCommentText(newText);
-            setTimeout(() => {
-                textarea.focus();
-                textarea.setSelectionRange(start + emoji.length, start + emoji.length);
-            }, 0);
-        } else {
-            setCommentText(prev => prev + emoji);
-        }
-        setShowEmojiPicker(false);
-    };
-
-    const handleBack = () => {
-        if (canGoBack) {
-            navigate(-1);
-        } else {
-            navigate('/feed');
-        }
-    };
-
-    const handleCopyLink = async () => {
-        const url = `${window.location.origin}/post/${postId}`;
-        try {
-            await navigator.clipboard.writeText(url);
-            setToast({ message: 'Ссылка скопирована!', type: 'success' });
-        } catch (err) {
-            setToast({ message: 'Не удалось скопировать ссылку', type: 'error' });
-        }
-        setShowMenu(false);
-    };
-
-    const handleEditPost = async (updatedData) => {
-        try {
-            const userData = getCookie('catsgram_user_data');
-            const userId = userData ? JSON.parse(userData).id : null;
-
-            await updatePost(postId, userId, updatedData.text, updatedData.imageFile);
-
-            setPost(prev => ({
-                ...prev,
-                text: updatedData.text,
-                imageUrl: updatedData.imageUrl || prev.imageUrl,
-                hasImage: updatedData.hasImage !== undefined ? updatedData.hasImage : prev.hasImage,
-            }));
-
-            setShowEditModal(false);
-            setToast({ message: 'Пост обновлён!', type: 'success' });
-        } catch (error) {
-            setToast({ message: 'Ошибка при обновлении поста', type: 'error' });
-        }
-    };
-
-    const handleDeletePost = async () => {
-        if (!window.confirm('Вы уверены что хотите удалить пост?')) return;
-
-        try {
-            await deletePost(postId);
-            setToast({ message: 'Пост удалён', type: 'success' });
-            setTimeout(() => navigate('/feed'), 1000);
-        } catch (error) {
-            setToast({ message: 'Ошибка при удалении поста', type: 'error' });
-        }
-    };
+        // Утилиты
+        formatCount,
+    } = usePostPage();
 
     if (loading) {
         return (
@@ -339,9 +76,7 @@ function PostPage() {
             <div className="post-page-new">
                 {/* Header */}
                 <div className="post-page-header">
-                    <button className="header-back-btn" onClick={handleBack}>
-                        ← Назад
-                    </button>
+                    <button className="header-back-btn" onClick={handleBack}>← Назад</button>
                     <h1 className="header-title">Пост</h1>
                     <div className="header-spacer"></div>
                 </div>
@@ -351,11 +86,10 @@ function PostPage() {
                     <div className="post-header-new">
                         <div className="post-author-new">
                             <div className="post-avatar-new">
-                                {post.avatar ? (
-                                    <img src={post.avatar} alt="avatar" />
-                                ) : (
-                                    <span>{post.username?.[0]?.toUpperCase() || '👤'}</span>
-                                )}
+                                {post.avatar
+                                    ? <img src={post.avatar} alt="avatar" />
+                                    : <span>{post.username?.[0]?.toUpperCase() || '👤'}</span>
+                                }
                             </div>
                             <div className="post-author-info-new">
                                 <div className="post-author-row">
@@ -366,12 +100,11 @@ function PostPage() {
                             </div>
                         </div>
 
-                        {/* Three Dots Menu */}
+                        {/* Меню */}
                         <div className="menu-container" ref={menuRef}>
                             <button className="menu-dots-btn" onClick={() => setShowMenu(!showMenu)}>
                                 ⋯
                             </button>
-
                             {showMenu && (
                                 <div className="menu-dropdown">
                                     <button className="menu-item" onClick={handleCopyLink}>
@@ -412,8 +145,8 @@ function PostPage() {
                                 ❤️ <span>{formatCount(post.likes)}</span>
                             </button>
                             <span className="stat-btn-new">
-                💬 <span>{formatCount(comments.length)}</span>
-              </span>
+                                💬 <span>{formatCount(comments.length)}</span>
+                            </span>
                         </div>
                     </div>
                 </div>
@@ -430,20 +163,19 @@ function PostPage() {
                             comments.map((comment) => (
                                 <div key={comment.id} className="comment-item-new">
                                     <div className="comment-avatar-new">
-                                        {comment.author?.avatar ? (
-                                            <img src={comment.author.avatar} alt="avatar" />
-                                        ) : (
-                                            <span>{comment.author?.username?.[0]?.toUpperCase() || '👤'}</span>
-                                        )}
+                                        {comment.author?.avatar
+                                            ? <img src={comment.author.avatar} alt="avatar" />
+                                            : <span>{comment.author?.username?.[0]?.toUpperCase() || '👤'}</span>
+                                        }
                                     </div>
                                     <div className="comment-content-new">
                                         <div className="comment-header-new">
-                                            <span className="comment-author-new">{comment.author?.username || `User${comment.authorId}`}</span>
+                                            <span className="comment-author-new">
+                                                {comment.author?.username || `User${comment.authorId}`}
+                                            </span>
                                             {comment.author?.verified && <span className="verified-badge-new">✅</span>}
                                             <span className="comment-time-new">{comment.time}</span>
                                         </div>
-
-                                        {/* Стало: */}
                                         <div className="comment-text-new">{parseMentions(comment.text)}</div>
                                     </div>
                                 </div>
@@ -456,7 +188,7 @@ function PostPage() {
                     </div>
                 </div>
 
-                {/* Comment Input - обнови эту секцию */}
+                {/* Comment Input */}
                 <div className="comment-input-container-new">
                     <div className="input-wrapper-comment">
                         <input
@@ -485,7 +217,6 @@ function PostPage() {
                         >
                             😊
                         </button>
-
                         {showEmojiPicker && (
                             <EmojiPicker
                                 onEmojiSelect={handleEmojiClick}
@@ -494,7 +225,6 @@ function PostPage() {
                             />
                         )}
                     </div>
-
                     <button
                         className="send-btn-new"
                         onClick={handleCommentSubmit}
@@ -506,7 +236,6 @@ function PostPage() {
                 </div>
             </div>
 
-            {/* Edit Post Modal */}
             {showEditModal && (
                 <EditPostModal
                     post={post}
@@ -520,16 +249,6 @@ function PostPage() {
             )}
         </Layout>
     );
-}
-
-function highlightMentions(text) {
-    if (!text) return '';
-    return text.split(/(@\w+)/g).map((part, index) => {
-        if (part.startsWith('@')) {
-            return <span key={index} className="mention-link">{part}</span>;
-        }
-        return part;
-    });
 }
 
 export default PostPage;
